requirements.txt
crewai[tools]>=0.80.0
crewai-tools>=0.14.0
duckduckgo-search>=6.0.0
ddgs>=0.0.1
langchain-community>=0.3.0
python-dotenv>=1.0.0
